apply PR review feedback
